Fix issues with the DB migrations and models
