--- data_models/job_type.py
@@ -1,10 +1,0 @@
-import enum
-
-
-class JobType(enum.Enum):
-    """
-    An enumeration representing the possible reduction states.
-    """
-    RERUN = "RERUN"
-    SIMPLE = "SIMPLE"
-    AUTOREDUCTION = "AUTOREDUCTION"
