--- data_models/state.py
@@ -1,12 +1,0 @@
-import enum
-
-
-class State(enum.Enum):
-    """
-    An enumeration representing the possible reduction states.
-    """
-
-    SUCCESSFUL = "SUCCESSFUL"
-    UNSUCCESSFUL = "UNSUCCESSFUL"
-    ERROR = "ERROR"
-    NOT_STARTED = "NOT_STARTED"
