+++ utils/__init__.py
@@ -1,0 +1,5 @@
+class DatabaseInconsistency(Exception):
+    """
+    Raised when the database is not consistent with expectations in code.
+    """
+    pass
